ML_Model/requirements.txt
fastapi==0.118.3
uvicorn[standard]==0.22.0
transformers==4.57.0
torch==2.6.0
torchvision==0.21.0
pillow==10.4.0
requests==2.31.0
numpy<2
